Se actualizo la lectura del .txt para poder utilzar PUSH y POP

+++ memory.h
@@ -2,7 +2,7 @@
 * \brief <center><b>Funcion Bitcount</b></center>
 * \param *Registro_activos para visualizar los registros activos
 */
-int Bitcount(int *R_activos);
+int Bitcount(uint8_t *R_activos);
 
 /** \fn void PUSH(uint32_t *Registro,uint8_t *Memory,int *R_activos);
 * \brief <center><b>Funcion PUSH</b></center>
@@ -10,7 +10,7 @@ int Bitcount(int *R_activos);
 * \param *Memory puntero dirigido a la memoria
 * \param *R_activos puntero a los resgistros activos
 */
-void PUSH(uint32_t *Registro,uint8_t *Memory,int *R_activos);
+void PUSH(uint32_t *Registro,uint8_t *Memory,uint8_t *R_activos);
 
 /** \fn void POP(uint32_t *Registro,uint8_t *Memory,int *R_activos);
 * \brief <center><b>Funcion POP</b></center>
@@ -18,4 +18,4 @@ void PUSH(uint32_t *Registro,uint8_t *Memory,int *R_activos);
 * \param *Memory puntero dirigido a la memoria
 * \param *R_activos puntero a los resgistros activos
 */
-void POP(uint32_t *Registro,uint8_t *Memory,int *R_activos);
+void POP(uint32_t *Registro,uint8_t *Memory,uint8_t *R_activos);
